I think I nailed the filesystem structure.
Created the base structure for file systems within the library. Everything is generalised; I'm only missing the implementations.

=== File Additions ===

* PermissionSet.java
Represents the permissions of a filesystem entry on a given device.
Internally uses values from the Permission enum (more on that further down)

=== File Mods ===

* ResourceManager.java
I don't think this will last long...
Ignore it for now.

* Permission.java
Scrapped the idea of using EnumSets for representing more complicated permissions.
All possible permissions are now in the Permissions enum.
Added parser methods to enum.

* IDirectory.java
Added all method prototypes.
This approach is in general a huge improvement over the original version of JDroidLib!

* IFile.java
Added all method prototypes.
This approach is in general a huge improvement over the original version of JDroidLib!
</copypasta>

* IFileSystemEntry.java
This interface is NOT intended to be implemented on its own to a class!
This interface is used ONLY for generalisations!
Implemented changes from Permissions enum.
Now uses PermissionSet.
Added method for getting the host(ing) device.

All in all, I think I nailed this one.