AndroidFile implementation almost complete.
Implementation of representation of files within the library is almost complete.

=== File Additions ===

* CannotTouchException.java
New device-class exception
Can/will be thrown when a file cannot be touched, or a directory cannot be created.

* FileCouldNotBePulledException.java
New device-class exception
Can/will be thrown when a file/directory could not be pulled to the host.

=== File Mods ===

* AndroidController.java
Added static method for getting instance of class or null
Handy for lib-internal stuff.
Don't recommend you use it in third-party applications. Might wreak havoc on your application.

* AndroidFile.java
Most of the new things are here
Added new properties for representing the hosting device and a string representation of the complete path of a file
Updated constructor to take Device argument (host device)
Error checking on host device
Accommodating for possibility of files possibly being hidden (period in front of name)
Accommodating for changes to underlying interfaces
Added pull functionality for files.
Added functionality for getting the contents of files (both as raw content and as strings)
Added functionality for touching (creating) files
Added functionality for determining whether a file is a text file or not
Added functionality for getting file permissions
Basic getter implementations
Added functionality for determining whether a file exists

* IDirectory.java
* IFile.java
Accommodating for changes to underlying interfaces

* IFileSystemEntry.java
Now contains "constants" defining useful command tags for use with these objects
Added exceptions to root methods
Added pull methods to this interface:
  No point in naming them differently in the specialisations. They essentially do the same thing.

* PermissionSet.java
Added way of getting an instance by providing output from `stat "%a %N" $FILE`
Just to make things cleaner in other classes.